Fixes as per Stefano's feedback
