+++ src/python/foglamp/translators/exceptions.py
