+++ src/sql/foglamp_init_data.sql
@@ -78,6 +78,9 @@ INSERT INTO foglamp.configuration ( key, description, value )
 INSERT INTO foglamp.configuration ( key, description, value )
      VALUES ( 'COAP', 'CoAP Plugin Configuration', ' { "plugin" : { "type" : "string", "value" : "coap", "default" : "coap", "description" : "Python module name of the plugin to load" } } ');
 
+INSERT INTO foglamp.configuration ( key, description, value )
+     VALUES ( 'HTTP_TR_3', 'HTTP North Plugin Configuration', ' { "plugin" : { "type" : "string", "value" : "http_translator", "default" : "http_translator", "description" : "Python module name of the plugin to load" } } ');
+
 -- DELETE data for roles, resources and permissions
 DELETE FROM foglamp.role_resource_permission;
 DELETE FROM foglamp.roles;
@@ -109,6 +112,7 @@ INSERT INTO foglamp.statistics ( key, description, value, previous_value )
             ( 'BUFFERED',   'The number of readings currently in the FogLAMP buffer', 0, 0 ),
             ( 'SENT_1',     'The number of readings sent to the historian', 0, 0 ),
             ( 'SENT_2',     'The number of statistics data sent to the historian', 0, 0 ),
+            ( 'SENT_3',     'The number of readings data sent to the HTTP translator', 0, 0 ),
             ( 'UNSENT',     'The number of readings filtered out in the send process', 0, 0 ),
             ( 'PURGED',     'The number of readings removed from the buffer by the purge process', 0, 0 ),
             ( 'UNSNPURGED', 'The number of readings that were purged from the buffer before being sent', 0, 0 ),
@@ -122,9 +126,12 @@ insert into foglamp.scheduled_processes (name, script) values ('COAP', '["python
 insert into foglamp.scheduled_processes (name, script) values ('purge', '["python3", "-m", "foglamp.data_purge"]');
 insert into foglamp.scheduled_processes (name, script) values ('stats collector', '["python3", "-m", "foglamp.statistics_history"]');
 insert into foglamp.scheduled_processes (name, script) values ('sending process', '["python3", "-m", "foglamp.sending_process", "--stream_id", "1", "--debug_level", "1"]');
+
 -- FogLAMP statistics into PI
 insert into foglamp.scheduled_processes (name, script) values ('statistics to pi','["python3", "-m", "foglamp.sending_process", "--stream_id", "2", "--debug_level", "1"]');
 
+-- send readings via HTTP
+insert into foglamp.scheduled_processes (name, script) values ('sending HTTP', '["python3", "-m", "foglamp.sending_process", "--stream_id", "3", "--debug_level", "1"]');
 
 -- Start the device server at start-up
 insert into foglamp.schedules(id, schedule_name, process_name, schedule_type,
@@ -150,15 +157,25 @@ schedule_time, schedule_interval, exclusive)
 values ('2b614d26-760f-11e7-b5a5-be2e44b06b34', 'sending process', 'sending process', 3,
 NULL, '00:00:15', true);
 
--- Run FogLAMP statistics into PI every 25 seconds
+---- Run FogLAMP statistics into PI every 25 seconds
 insert into foglamp.schedules(id, schedule_name, process_name, schedule_type,
 schedule_time, schedule_interval, exclusive)
 values ('1d7c327e-7dae-11e7-bb31-be2e44b06b34', 'statistics to pi', 'statistics to pi', 3,
 NULL, '00:00:25', true);
 
+-- Run the sending process using HTTP translator every 15 seconds
+insert into foglamp.schedules(id, schedule_name, process_name, schedule_type,
+schedule_time, schedule_interval, exclusive)
+values ('81bdf749-8aa0-468e-b229-9ff695668e8c', 'sending via HTTP', 'sending HTTP', 3,
+NULL, '00:00:15', true);
+
 -- OMF translator configuration
-INSERT INTO foglamp.destinations(id,description, ts)                       VALUES (1,'OMF', now());
+INSERT INTO foglamp.destinations(id,description, ts) VALUES (1,'OMF', now());
 INSERT INTO foglamp.streams(id,destination_id,description, last_object,ts) VALUES (1,1,'OMF translator', 0,now());  
 
 -- FogLAMP statistics into PI configuration
 INSERT INTO foglamp.streams (id,destination_id,description, last_object,ts ) VALUES (2,1,'FogLAMP statistics into PI', 0,now());
+
+-- HTTP translator configuration
+INSERT INTO foglamp.destinations(id,description, ts) VALUES (2,'HTTP_TR', now());
+INSERT INTO foglamp.streams(id,destination_id,description, last_object,ts) VALUES (3,2,'HTTP translator', 0,now());
